Add scrape scripts of top charts (Radio or Spotify)
For scripts that gather national top charts from radio or top list
sources, the files are named 'scrape_top_charts_NATION.py'. Each of
these scripts will create a separate dataset. All scandinavian countries
except Iceland are present. The schema of the datasets is
'position,artist,title,year,week'.

For Spotify, the file is 'scrape_top_charts_spotify.py'
and creates a single dataset with all scandinavian countries. The schema
of the dataset is 'position,artist,title,streams,daterange,year,country'

The 'detect_language.py' script takes in any dataset of the above
mentioned and creates a new 'language' column with the language of the
song in that row. A Genius access token is needed to make this work,
refer to https://lyricsgenius.readthedocs.io/en/master/setup.html