Add scrape scripts and labelled dataset of Eurovision Song Contest
The official ESC website https://eurovision.tv/ is taken as reference
for the list of all participants in the history of the contest. The
website https://www.esc-history.com/ offers information about language
of the song. After integrating the two sources, 606 songs are still left
without language information. The final dataset is manually labeled to
include the language of every song presented at Eurovision.