Label filtered datasets with language column
A script is developed to add the `language` column to the previously
filtered datasets. The language of a song is detected through the
`langdetect` Python module. The input to the detection algorithm is a
string, either the full lyrics of the song taken from the
https://genius.com/ website if available or just the song title.

Obtained labelled datasets are stored in a separate folder.